+++ docs/changes.rst
@@ -172,6 +172,15 @@ Released: not yet
 
 **Deprecations:**
 
+  Deprecated the method FakedWBEMConnection.compile_dmtf_schema in favor of
+  a new method FakedWBEMConnection.compile_dmtf_classes that does not
+  directly use the DMTFCIMSchema class.
+
+  Deprecated the attribute name schema_mof_file in DMTFCIMSchema in favor of
+  the name schema_pragma_file since this is the file that contains all
+  of the MOF pragmas defining the locations of the class MOF files in a
+  set of directories.
+
 **Bug fixes:**
 
 * Dev/Test: Pinned lxml to <4.4.0 because that version removed Python 3.4
@@ -616,6 +625,9 @@ Released: not yet
   CreateInstance method defined in the compiler repo.  For the tests that
   means that the path creation is in MOFWBEMConnection.CreateInstance.
   (See issue # 1911)
+
+* Added a new method FakedWBEMConnection.compile_dmtf_classes that does not
+  directly use the DMTFCIMSchema class.  See the Deprecations section.
 
 **Known issues:**